refactor(store): move room store types under app/lib/store
